test: move RestDocsXXX to restdocs.support package
