+++ src/test/java/com/meoguri/linkocean/test/restdocs/BookmarkRestDocsTest.java
@@ -20,7 +20,7 @@ import org.springframework.test.web.servlet.ResultActions;
 import com.meoguri.linkocean.controller.bookmark.BookmarkController;
 import com.meoguri.linkocean.controller.bookmark.dto.RegisterBookmarkRequest;
 import com.meoguri.linkocean.controller.bookmark.dto.UpdateBookmarkRequest;
-import com.meoguri.linkocean.test.support.controller.RestDocsTestSupport;
+import com.meoguri.linkocean.test.restdocs.support.RestDocsTestSupport;
 
 class BookmarkRestDocsTest extends RestDocsTestSupport {
