+++ src/test/java/com/meoguri/linkocean/test/restdocs/FollowRestDocsTest.java
@@ -14,7 +14,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.test.web.servlet.ResultActions;
 
 import com.meoguri.linkocean.controller.profile.FollowController;
-import com.meoguri.linkocean.test.support.controller.RestDocsTestSupport;
+import com.meoguri.linkocean.test.restdocs.support.RestDocsTestSupport;
 
 class FollowRestDocsTest extends RestDocsTestSupport {
