+++ src/test/java/com/meoguri/linkocean/test/restdocs/LoginRestDocsTest.java
@@ -14,7 +14,7 @@ import org.springframework.test.web.servlet.ResultActions;
 
 import com.meoguri.linkocean.controller.user.LoginController;
 import com.meoguri.linkocean.controller.user.dto.LoginRequest;
-import com.meoguri.linkocean.test.support.controller.RestDocsTestSupport;
+import com.meoguri.linkocean.test.restdocs.support.RestDocsTestSupport;
 
 class LoginRestDocsTest extends RestDocsTestSupport {
