+++ src/test/java/com/meoguri/linkocean/test/restdocs/ReactionRestDocsTest.java
@@ -14,7 +14,7 @@ import org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders;
 import org.springframework.test.web.servlet.ResultActions;
 
 import com.meoguri.linkocean.controller.bookmark.ReactionController;
-import com.meoguri.linkocean.test.support.controller.RestDocsTestSupport;
+import com.meoguri.linkocean.test.restdocs.support.RestDocsTestSupport;
 
 class ReactionRestDocsTest extends RestDocsTestSupport {
