+++ src/test/java/com/meoguri/linkocean/test/restdocs/support/RestDocsTestSupport.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.test.support.controller;
+package com.meoguri.linkocean.test.restdocs.support;
 
 import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.*;
 
@@ -10,6 +10,8 @@ import org.springframework.restdocs.mockmvc.RestDocumentationResultHandler;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 import org.springframework.web.filter.CharacterEncodingFilter;
+
+import com.meoguri.linkocean.test.support.controller.BaseControllerTest;
 
 @RestDocsTest
 public abstract class RestDocsTestSupport extends BaseControllerTest {
